repositories_used/system-4b-augmentation-phase-1/requirements.txt
blend_modes==2.1.0
dlib==19.24.0
matplotlib==3.5.1
mediapipe==0.8.9.1
numpy==1.21.5
onnxruntime==1.12.1
opencv_contrib_python==4.5.5.64
pandas==1.3.5
Pillow==9.2.0
protobuf3_to_dict==0.1.5
protobuf_to_dict==0.1.0
pyshtools==4.10.1
pytorchcv==0.0.67
scikit_image==0.19.3
scikit_learn==1.1.2
scipy==1.6.2
skimage==0.0
torch==1.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.0
